task list commit

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,28 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+                ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else{
+            return view('welcome');
+        }
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+     /**   
+      * $tasks = Task::all();
+       * 
+        * return view('tasks.index', [
+         *   'tasks' => $tasks,
+        *    ]);
+        */
     }
 
     /**
@@ -49,12 +66,13 @@ class TasksController extends Controller
                 'content' => 'required|max:191',
             ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
-        return redirect('/');
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+            
+            ]);
+
+        return redirect()->back();
     }
 
     /**
@@ -117,9 +135,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
         
-        return redirect('/');
+        if(\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back;
     }
 }
